requirements.txt
python-dotenv==1.2.1
requests==2.32.5
requests-oauthlib==2.0.0
cloudinary
Pillow
anthropic
pytz
duckduckgo-search
